revert source files to release-2.26.3

+++ src/haptic/windows/SDL_windowshaptic_c.h
@@ -18,11 +18,12 @@
      misrepresented as being the original software.
   3. This notice may not be removed or altered from any source distribution.
 */
-#include "SDL_internal.h"
+#include "../../SDL_internal.h"
 
 #ifndef SDL_windowshaptic_c_h_
 #define SDL_windowshaptic_c_h_
 
+#include "SDL_thread.h"
 #include "../SDL_syshaptic.h"
 #include "../../core/windows/SDL_directx.h"
 #include "../../core/windows/SDL_xinput.h"
@@ -40,15 +41,16 @@ struct haptic_hwdata
 #if SDL_HAPTIC_DINPUT
     LPDIRECTINPUTDEVICE8 device;
 #endif
-    DWORD axes[3];        /* Axes to use. */
-    SDL_bool is_joystick; /* Device is loaded as joystick. */
-    Uint8 bXInputHaptic;  /* Supports force feedback via XInput. */
-    Uint8 userid;         /* XInput userid index for this joystick */
+    DWORD axes[3];              /* Axes to use. */
+    SDL_bool is_joystick;       /* Device is loaded as joystick. */
+    Uint8 bXInputHaptic; /* Supports force feedback via XInput. */
+    Uint8 userid; /* XInput userid index for this joystick */
     SDL_Thread *thread;
     SDL_mutex *mutex;
-    Uint64 stopTicks;
+    Uint32 stopTicks;
     SDL_atomic_t stopThread;
 };
+
 
 /*
  * Haptic system effect data.
@@ -67,8 +69,8 @@ struct haptic_hweffect
 #endif
 
 /*
- * List of available haptic devices.
- */
+* List of available haptic devices.
+*/
 typedef struct SDL_hapticlist_item
 {
     char *name;
@@ -78,7 +80,7 @@ typedef struct SDL_hapticlist_item
     DIDEVCAPS capabilities;
 #endif
     SDL_bool bXInputHaptic; /* Supports force feedback via XInput. */
-    Uint8 userid;           /* XInput userid index for this joystick */
+    Uint8 userid; /* XInput userid index for this joystick */
     struct SDL_hapticlist_item *next;
 } SDL_hapticlist_item;
 
@@ -93,3 +95,6 @@ extern int SDL_SYS_RemoveHapticDevice(SDL_hapticlist_item *prev, SDL_hapticlist_
 #endif
 
 #endif /* SDL_windowshaptic_c_h_ */
+
+/* vi: set ts=4 sw=4 expandtab: */
+
